Renamed lcsources as unifieddownloader
